improve api call function names
