+++ my-app/api/user/get-user-details.ts
@@ -2,20 +2,10 @@ import { User } from "@/interfaces/user";
 import { ApiResponse } from "@/interfaces/api-response";
 import { apiClient } from "@/api";
 import { AxiosError, isAxiosError } from "axios";
-import { PossibleRolesEn } from "@/utils";
 
-interface AddUserRequest {
-  name: string;
-  email: string;
-  phone: string;
-  birthDate: string;
-  password: string;
-  role: PossibleRolesEn;
-}
-
-export async function add(request: AddUserRequest): Promise<ApiResponse<User>> {
+export async function getUserDetails(id: string): Promise<ApiResponse<User>> {
   const response = await apiClient
-    .post<ApiResponse<User>>("/users", request)
+    .get<ApiResponse<User>>(`/users/${id}`)
     .catch((error: Error | AxiosError) => {
       if (isAxiosError(error)) {
         return error.response;
